+++ Week_01/id_31/LeetCode_1047_31.cpp
